+++ gson/src/main/java/org/baeldung/gson/primitives/charvalue/JsonUnicodeCharValue.java
@@ -1,15 +1,14 @@
-package com.baeldung.gson_primitive_types.char_value;
+package org.baeldung.gson.primitives.charvalue;
 
-import com.baeldung.gson_primitive_types.models.GsonBitString;
-import com.baeldung.gson_primitive_types.models.GsonLatinChar;
 import com.google.gson.Gson;
+import org.baeldung.gson.primitives.models.GsonChar;
 
 public class JsonUnicodeCharValue {
     public static void main(String[] args) {
         // The field is converted.
         Gson gson = new Gson();
         String json = "{\"value\": \"\\u00AE\"}";
-        GsonLatinChar model = gson.fromJson(json, GsonLatinChar.class);
+        GsonChar model = gson.fromJson(json, GsonChar.class);
 
         System.out.println(model);
     }
