+++ gson/src/main/java/org/baeldung/gson/primitives/models/GsonBitString.java
@@ -1,4 +1,4 @@
-package com.baeldung.gson_primitive_types.models;
+package org.baeldung.gson.primitives.models;
 
 public class GsonBitString {
     public byte value = (byte) 1;
