+++ gson/src/main/java/org/baeldung/gson/primitives/specialvalue/JsonEmptyValue.java
@@ -1,13 +1,13 @@
-package com.baeldung.gson_primitive_types.special_value;
+package org.baeldung.gson.primitives.specialvalue;
 
-import com.baeldung.gson_primitive_types.models.GsonBitString;
 import com.google.gson.Gson;
+import org.baeldung.gson.primitives.models.GsonBitString;
 
-public class JsonInvalidStringValue {
+public class JsonEmptyValue {
     public static void main(String[] args) {
         // Raises an exception.
         Gson gson = new Gson();
-        String json = "{\"value\": \"15x\"}";
+        String json = "{\"value\": \"\"}";
         GsonBitString model = gson.fromJson(json, GsonBitString.class);
 
         System.out.println(model);
