+++ gson/src/main/java/org/baeldung/gson/primitives/specialvalue/JsonInvalidStringValue.java
@@ -1,13 +1,13 @@
-package com.baeldung.gson_primitive_types.special_value;
+package org.baeldung.gson.primitives.specialvalue;
 
-import com.baeldung.gson_primitive_types.models.GsonBitString;
 import com.google.gson.Gson;
+import org.baeldung.gson.primitives.models.GsonBitString;
 
-public class JsonInvalidValue {
+public class JsonInvalidStringValue {
     public static void main(String[] args) {
         // Raises an exception.
         Gson gson = new Gson();
-        String json = "{\"value\": s15s}";
+        String json = "{\"value\": \"15x\"}";
         GsonBitString model = gson.fromJson(json, GsonBitString.class);
 
         System.out.println(model);
