+++ gson/src/main/java/org/baeldung/gson/primitives/specialvalue/JsonNullValue.java
@@ -1,7 +1,7 @@
-package com.baeldung.gson_primitive_types.special_value;
+package org.baeldung.gson.primitives.specialvalue;
 
-import com.baeldung.gson_primitive_types.models.GsonBitString;
 import com.google.gson.Gson;
+import org.baeldung.gson.primitives.models.GsonBitString;
 
 public class JsonNullValue {
     public static void main(String[] args) {
